Replace original approach
