+++ packages/components/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Introduce `Navigation` component as `__experimentalNavigation` for displaying a heirarchy of items.
+- Introduce `Navigation` component as `__experimentalNavigation` for displaying a hierarchy of items.
 
 ## 10.0.0 (2020-07-07)
