+++ packages/components/src/index.js
@@ -67,11 +67,10 @@ export { default as MenuItemsChoice } from './menu-items-choice';
 export { default as Modal } from './modal';
 export { default as ScrollLock } from './scroll-lock';
 export { NavigableMenu, TabbableContainer } from './navigable-container';
-export {
-	default as __experimentalNavigation,
-	NavigationMenu as __experimentalNavigationMenu,
-	NavigationMenuItem as __experimentalNavigationMenuItem,
-} from './navigation';
+export { default as __experimentalNavigation } from './navigation';
+export { default as __experimentalNavigationGroup } from './navigation/group';
+export { default as __experimentalNavigationItem } from './navigation/item';
+export { default as __experimentalNavigationLevel } from './navigation/level';
 export { default as Notice } from './notice';
 export { default as __experimentalNumberControl } from './number-control';
 export { default as NoticeList } from './notice/list';
