+++ packages/components/src/navigation/index.js
@@ -2,126 +2,86 @@
  * External dependencies
  */
 import classnames from 'classnames';
+import { noop } from 'lodash';
 
 /**
  * WordPress dependencies
  */
-import { usePrevious } from '@wordpress/compose';
-import { useEffect, useMemo, useState, useRef } from '@wordpress/element';
-import { Icon, chevronLeft } from '@wordpress/icons';
+import { useEffect, useRef, useState } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
 import Animate from '../animate';
-import { BackButtonUI, Root } from './styles/navigation-styles';
+import { NavigationContext } from './context';
+import { DEFAULT_LEVEL } from './utils';
+import { Root } from './styles/navigation-styles';
 
-const Navigation = ( { activeItemId, children, data, rootTitle } ) => {
-	const [ activeLevelId, setActiveLevelId ] = useState( 'root' );
+export default function Navigation( {
+	children,
+	activeItem,
+	activeLevel = DEFAULT_LEVEL,
+	setActiveItem = noop,
+	setActiveLevel = noop,
+} ) {
+	const [ item, setItem ] = useState( activeItem );
+	const [ level, setLevel ] = useState( activeLevel );
+	const [ slideOrigin, setSlideOrigin ] = useState();
 
-	const appendItemData = ( item ) => {
-		return {
-			...item,
-			children: [],
-			parent: item.id === 'root' ? null : item.parent || 'root',
-			isActive: item.id === activeItemId,
-			setActiveLevelId,
-		};
+	const activateItem = ( itemId ) => {
+		setItem( itemId );
+		setActiveItem( itemId );
 	};
 
-	const mapItems = ( itemData ) => {
-		const items = new Map(
-			[
-				{ id: 'root', parent: null, title: rootTitle },
-				...itemData,
-			].map( ( item ) => [ item.id, appendItemData( item ) ] )
-		);
-
-		items.forEach( ( item ) => {
-			const parentItem = items.get( item.parent );
-			if ( parentItem ) {
-				parentItem.children.push( item );
-				parentItem.hasChildren = true;
-			}
-		} );
-
-		return items;
+	const activateLevel = ( levelId, slideInOrigin = 'left' ) => {
+		setSlideOrigin( slideInOrigin );
+		setLevel( levelId );
+		setActiveLevel( levelId );
 	};
-
-	const items = useMemo( () => mapItems( data ), [
-		data,
-		activeItemId,
-		rootTitle,
-	] );
-	const activeItem = items.get( activeItemId );
-	const previousActiveLevelId = usePrevious( activeLevelId );
-	const level = items.get( activeLevelId );
-	const parentLevel = level && items.get( level.parent );
-	const isNavigatingBack =
-		previousActiveLevelId &&
-		items.get( previousActiveLevelId ).parent === activeLevelId;
-
-	useEffect( () => {
-		if ( activeItem ) {
-			setActiveLevelId( activeItem.parent );
-		}
-	}, [ activeItem ] );
 
 	const isMounted = useRef( false );
-
 	useEffect( () => {
 		if ( ! isMounted.current ) {
 			isMounted.current = true;
 		}
 	}, [] );
 
-	const NavigationBackButton = ( { children: backButtonChildren } ) => {
-		if ( ! parentLevel ) {
-			return null;
+	useEffect( () => {
+		if ( activeItem !== item ) {
+			activateItem( activeItem );
 		}
+		if ( activeLevel !== level ) {
+			activateLevel( activeLevel );
+		}
+	}, [ activeItem, activeLevel ] );
 
-		return (
-			<BackButtonUI
-				className="components-navigation__back-button"
-				isTertiary
-				onClick={ () => setActiveLevelId( parentLevel.id ) }
-			>
-				<Icon icon={ chevronLeft } />
-				{ backButtonChildren }
-			</BackButtonUI>
-		);
+	const context = {
+		activeItem: item,
+		activeLevel: level,
+		setActiveItem: activateItem,
+		setActiveLevel: activateLevel,
 	};
 
 	return (
 		<Root className="components-navigation">
 			<Animate
-				key={ level.id }
+				key={ level }
 				type="slide-in"
-				options={ {
-					origin: isNavigatingBack ? 'right' : 'left',
-				} }
+				options={ { origin: slideOrigin } }
 			>
 				{ ( { className: animateClassName } ) => (
 					<div
-						className={ classnames(
-							'components-navigation__level',
-							{
-								[ animateClassName ]: isMounted.current,
-							}
-						) }
-					>
-						{ children( {
-							level,
-							NavigationBackButton,
-							parentLevel,
+						className={ classnames( {
+							[ animateClassName ]:
+								isMounted.current && slideOrigin,
 						} ) }
+					>
+						<NavigationContext.Provider value={ context }>
+							{ children }
+						</NavigationContext.Provider>
 					</div>
 				) }
 			</Animate>
 		</Root>
 	);
-};
-
-export default Navigation;
-export { default as NavigationMenu } from './menu';
-export { default as NavigationMenuItem } from './menu-item';
+}
