+++ packages/components/src/navigation/utils.js
@@ -1,0 +1,1 @@
+export const DEFAULT_LEVEL = 'root';
